requirements.txt
tensorflow
numpy
pandas
PILasOPENCV
langchain-groq
python-dotenv